fix: eliminate all clippy warnings (v15.2.0)
Code quality maintenance - fixed 20 clippy warnings to 0:

Style fixes:
- proactive_alerts.rs: rename from_str to parse_from_str (avoid std trait confusion)
- proactive_alerts.rs: use or_default() instead of or_insert_with
- routes.rs: use is_none_or() instead of map_or()
- app.rs: remove unnecessary u64 cast
- components.rs: use .to_string() instead of format!()

Dead code annotations (intentionally defined for future use):
- proactive_alerts.rs: add #[allow(dead_code)] to AlertSeverity, AlertCategory,
  ProactiveAlert, AlertRule, ProactiveAlertStats, ProactiveAlertManager and all methods
- tool_sequence_advisor.rs: add #[allow(dead_code)] to based_on_pattern field
- routes.rs: add #[allow(dead_code)] to tool_sequence_advisor field in HttpState
- state.rs: add #[allow(dead_code)] to id and acknowledged fields in ProactiveAlertDisplay

cargo clippy 0 warnings
cargo test 506 tests passed

+++ src/tui/app.rs
@@ -713,7 +713,7 @@ impl TuiApp {
                 use std::time::SystemTime;
 
                 // Format the timestamp
-                let created_str = match SystemTime::UNIX_EPOCH.checked_add(std::time::Duration::from_secs(created_at as u64)) {
+                let created_str = match SystemTime::UNIX_EPOCH.checked_add(std::time::Duration::from_secs(created_at)) {
                     Some(time) => {
                         let datetime: chrono::DateTime<chrono::Utc> = time.into();
                         datetime.format("%H:%M:%S").to_string()
